Going through the Network Core and resolving a bunch of compiler issues that were introduced by combining file-rewrites into the existing files. Namely lots of redeclarations, and syntax errors introduced by using Powershell to combine files together.
We are rapidly approaching being able to actually compile a final binary to test properly.

Once compiler issues are all resolved, we will finish the testing suite. Once that is all done, we are ready to move on to compiling the finished binary and testing. From there, we are ready for alpha release. After several months of work, we are very near ready for a public alpha release